requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
python-dotenv==1.0.0
scispacy==0.5.3
spacy>=3.6.0,<3.7.0
pydantic==2.5.0
python-multipart==0.0.6
numpy<2.0.0